+++ platforms/gradle/better-gradle-properties/src/main/kotlin/foundry/gradle/properties/PropertyResolver.kt
@@ -111,16 +111,53 @@ public class PropertyResolver(
       ?: error("No property for '$key' found and no default value was provided.")
   }
 
+  public fun requiredStringProvider(key: String): Provider<String> {
+    return optionalStringProvider(key)
+      .orElse(
+        providers.provider {
+          error("No property for '$key' found and no default value was provided.")
+        }
+      )
+  }
+
   public fun optionalStringValue(key: String, defaultValue: String? = null): String? {
     return providerFor(key).orNull ?: defaultValue
   }
 
-  public fun optionalStringProvider(key: String): Provider<String> {
-    return optionalStringProvider(key, null)
+  public fun optionalStringProvider(
+    key: String,
+    blankBehavior: BlankBehavior = BlankBehavior.ERROR,
+  ): Provider<String> {
+    return optionalStringProvider(key, null, blankBehavior)
   }
 
-  public fun optionalStringProvider(key: String, defaultValue: String? = null): Provider<String> {
-    return providerFor(key).let { defaultValue?.let { providers.provider { defaultValue } } ?: it }
+  public fun optionalStringProvider(
+    key: String,
+    defaultValue: String? = null,
+    blankBehavior: BlankBehavior = BlankBehavior.ERROR,
+  ): Provider<String> {
+    return providerFor(key)
+      .let { defaultValue?.let { providers.provider { defaultValue } } ?: it }
+      .filter {
+        when (blankBehavior) {
+          BlankBehavior.FILTER -> {
+            it.isNotBlank()
+          }
+          BlankBehavior.ERROR -> {
+            it.ifBlank { error("Unexpected blank value for property '$key'") }
+            true
+          }
+          BlankBehavior.NONE -> {
+            true
+          }
+        }
+      }
+  }
+
+  public enum class BlankBehavior {
+    FILTER,
+    ERROR,
+    NONE,
   }
 
   public companion object {